requirements.txt
streamlit
tensorflow
tf-explain
opencv-python
matplotlib
pillow
